refactor: github 관련 코드 auth 패키지로 이동
