+++ src/test/java/nextstep/auth/acceptance/GithubResponses.java
@@ -1,4 +1,4 @@
-package nextstep.member.acceptance;
+package nextstep.auth.acceptance;
 
 import java.util.Arrays;
